+++ kmep_harness/android/app/src/main/kotlin/com/example/kmep_harness/MainActivity.kt
@@ -1,4 +1,4 @@
-package com.example.kmep_proto_harness
+package com.example.kmep_harness
 
 import io.flutter.embedding.android.FlutterActivity